QuantumNumbers submodule
