+++ src/lib/ITensorMPS/src/opsum_to_mpo/opsum_to_mpo_generic.jl
@@ -1,4 +1,5 @@
 using NDTensors: using_auto_fermion
+using ..Ops: Op
 
 # TODO: Deprecate.
 const AutoMPO = OpSum
